Bring EM problem depreciation to all problems.
